Removed unnecessary conversion

+++ src/main/scala/ch/epfl/bluebrain/nexus/admin/index/OrganizationCache.scala
@@ -52,7 +52,7 @@ class OrganizationCache[F[_]](store: KeyValueStore[F, UUID, OrganizationResource
             acls.exists(organization.label, orgs.read)
       }
       val count  = filtered.size.toLong
-      val result = filtered.toList.sorted.slice(pagination.from.toInt, (pagination.from + pagination.size).toInt)
+      val result = filtered.toList.sorted.slice(pagination.from, (pagination.from + pagination.size))
       UnscoredQueryResults(count, result.map(UnscoredQueryResult(_)))
     }
